# openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/upload/presign": {
      "post": {
        "operationId": "UploadController_presign",
        "summary": "Generate one-shot upload URL",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PresignRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresignedUrlResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "upload"
        ]
      }
    },
    "/upload/complete": {
      "post": {
        "operationId": "UploadController_complete",
        "summary": "Confirm one-shot upload",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmUploadDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfirmUploadResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "upload"
        ]
      }
    },
    "/upload/multipart/start": {
      "post": {
        "operationId": "UploadController_startMultipartUpload",
        "summary": "Start multipart upload session",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartMultipartUploadDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MultipartUploadSessionResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "upload"
        ]
      }
    },
    "/upload/multipart/part": {
      "post": {
        "operationId": "UploadController_presignMultipartPart",
        "summary": "Presign multipart upload part",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PresignMultipartPartDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MultipartUploadPartResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "upload"
        ]
      }
    },
    "/upload/multipart/complete": {
      "post": {
        "operationId": "UploadController_completeMultipartUpload",
        "summary": "Complete multipart upload session",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteMultipartUploadDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MultipartUploadCompleteResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "upload"
        ]
      }
    },
    "/upload/multipart/abort": {
      "post": {
        "operationId": "UploadController_abortMultipartUpload",
        "summary": "Abort multipart upload session",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AbortMultipartUploadDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MultipartUploadAbortResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "upload"
        ]
      }
    },
...
  },
  "info": {
    "title": "Interview App API",
    "description": "HTTP API contract for Interview App backend",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session"
      }
    },
    "schemas": {
...
      "ApiErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400
          },
          "error": {
            "type": "string",
            "example": "Bad Request"
          },
          "message": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ],
            "example": "Validation failed"
          },
          "path": {
            "type": "string",
            "example": "/questions/invalid-id"
          }
        },
        "required": [
          "statusCode",
          "error",
          "message"
        ]
      },
...
      "PresignRequestDto": {
        "type": "object",
        "properties": {
          "questionIndex": {
            "type": "number"
          },
          "contentType": {
            "type": "string",
            "enum": [
              "video/webm"
            ]
          },
          "mediaType": {
            "type": "string",
            "enum": [
              "camera",
              "screen"
            ]
          }
        },
        "required": [
          "questionIndex",
          "contentType"
        ]
      },
      "PresignedUrlResponseDto": {
        "type": "object",
        "properties": {
          "uploadUrl": {
            "type": "string"
          },
          "mediaKey": {
            "type": "string"
          }
        },
        "required": [
          "uploadUrl",
          "mediaKey"
        ]
      },
      "ConfirmUploadDto": {
        "type": "object",
        "properties": {
          "questionIndex": {
            "type": "number"
          },
          "mediaKey": {
            "type": "string"
          }
        },
        "required": [
          "questionIndex",
          "mediaKey"
        ]
      },
      "ConfirmUploadResponseDto": {
        "type": "object",
        "properties": {
          "mediaKey": {
            "type": "string"
          },
          "confirmed": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "mediaKey",
          "confirmed"
        ]
      },
      "StartMultipartUploadDto": {
        "type": "object",
        "properties": {
          "questionIndex": {
            "type": "number"
          },
          "contentType": {
            "type": "string",
            "enum": [
              "video/webm"
            ]
          },
          "mediaType": {
            "type": "string",
            "enum": [
              "camera",
              "screen"
            ]
          }
        },
        "required": [
          "questionIndex",
          "contentType"
        ]
      },
      "MultipartUploadSessionResponseDto": {
        "type": "object",
        "properties": {
          "mediaKey": {
            "type": "string"
          },
          "uploadId": {
            "type": "string"
          }
        },
        "required": [
          "mediaKey",
          "uploadId"
        ]
      },
      "PresignMultipartPartDto": {
        "type": "object",
        "properties": {
          "questionIndex": {
            "type": "number"
          },
          "mediaKey": {
            "type": "string"
          },
          "uploadId": {
            "type": "string"
          },
          "partNumber": {
            "type": "number"
          }
        },
        "required": [
          "questionIndex",
          "mediaKey",
          "uploadId",
          "partNumber"
        ]
      },
      "MultipartUploadPartResponseDto": {
        "type": "object",
        "properties": {
          "mediaKey": {
            "type": "string"
          },
          "uploadId": {
            "type": "string"
          },
          "partNumber": {
            "type": "number"
          },
          "uploadUrl": {
            "type": "string"
          }
        },
        "required": [
          "mediaKey",
          "uploadId",
          "partNumber",
          "uploadUrl"
        ]
      },
      "CompleteMultipartUploadDto": {
        "type": "object",
        "properties": {
          "questionIndex": {
            "type": "number"
          },
          "mediaKey": {
            "type": "string"
          },
          "uploadId": {
            "type": "string"
          }
        },
        "required": [
          "questionIndex",
          "mediaKey",
          "uploadId"
        ]
      },
      "MultipartUploadCompleteResponseDto": {
        "type": "object",
        "properties": {
          "mediaKey": {
            "type": "string"
          },
          "uploadId": {
            "type": "string"
          },
          "completed": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "mediaKey",
          "uploadId",
          "completed"
        ]
      },
      "AbortMultipartUploadDto": {
        "type": "object",
        "properties": {
          "questionIndex": {
            "type": "number"
          },
          "mediaKey": {
            "type": "string"
          },
          "uploadId": {
            "type": "string"
          }
        },
        "required": [
          "questionIndex",
          "mediaKey",
          "uploadId"
        ]
      },
      "MultipartUploadAbortResponseDto": {
        "type": "object",
        "properties": {
          "mediaKey": {
            "type": "string"
          },
          "uploadId": {
            "type": "string"
          },
          "aborted": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "mediaKey",
          "uploadId",
          "aborted"
        ]
      },
...
    }
  }
}